refactor comscoreAnalytics unit tests
